internal/k8s/statefulset.go
package k8s

import (
	"context"

	"k8s.io/apimachinery/pkg/api/errors"
	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
	"k8s.io/client-go/kubernetes"
)

// HasStatefulSet checks if a StatefulSet with the given name exists in the specified namespace.
func HasStatefulSet(clientset kubernetes.Interface, namespace, name string) (bool, error) {
	_, err := clientset.AppsV1().StatefulSets(namespace).Get(context.TODO(), name, metav1.GetOptions{})
	if err == nil {
		return true, nil
	}
	if errors.IsNotFound(err) {
		return false, nil
	}
	return false, err
}
